requirements.txt
pandas
requests
tqdm
azure-identity